initial support for execution of queries
The query logic is entirely abstracted in InvenioQuery.  This will allow (with
some light refactoring) supporting backends other than tracker.  The search
window will execute the query and then update the results view.

The results are updated in place in the results view.  If there are matches from
the previous query, the existing matches are first overwritten.  If there are
more matches than previously, the new matches are inserted.  If there are fewer
matches, then the remaining matches are removed.