drop unnecessary subclassing of GTK widgets
The only thing that we are gaining by subclassing GtkStatusIcon and
GtkWindow is the dispose method.  Because these are eternal singletons,
there is very little point in spending the effort to subclass them.  We
can add free functions or just let the memory be cleaned up anyways when
we exit (which is when the classes will be destroyed).