add support for alternate actions
On Ctrl-Enter, we perform the alternate action (opening the containing
directory).  The nie:url(location) did not function as expected, and
required the conversion to an inline query.

Minor refactoring to remove the row-activated signal.  We end up using
it to call another function.  Just avoid the signal entirely and call
the function ourseleves.