statically define the column types with association
Create a static const table mapping the column to a particular GType.
The compiler should be smart enough to realise that the macro just does
a static lookup into the array and inline that, and then realise that
the value never changes and inline that.  This means that there should
be 0 overhead for this trick with optimisations in text or runtime
(though this is run once, so the runtime cost is negligible).

This helps provide some safety in the future though as the types are not
hardcoded for the list store, only the fields.  If we update the fields
in the future, the compiler will notify us that we are using a
non-existent field.