update TODO with ideas for window movement fixes
Override Redirect is one option for ensuring that the window cannot be moved.
However, it moves the window entirely out of the window manager's control.  We
need to implement custom handlers for all events, and manually grab keyboard and
pointer focus.

SubstructureRedirect is another option.  By setting the
SubstructureRedirectMask, we can get ConfigureRequest events, which if ignored
from GDK's stream would allow us to prevent the window from being moved.