update TODO
Finally figured out abustany's window placement bug.  We use gtk_window_move to
move the window to the proper location.  However, window managers are free to
ignore this call.  The result is that the search window can be displayed in an
incorrect location.

Set metacity's disable_workarounds option to reproduce this bug.