update TODO
We dont want to use OverrideRedirect with this window nor GTK_WINDOW_POPUP.  We
want the window to remain under the control of the window manager as we want to
be able to freely move the mouse and change focus to other windows.  The intent
is to simply prevent the move of the window similar to a menu.  Fortunately, we
can simply filter ConfigureRequest from ever reaching GDK and prevent that.